+++ backend/src/main/java/com/ody/auth/service/apple/AppleRevokeTokenClient.java
@@ -1,6 +1,7 @@
-package com.ody.auth.service;
+package com.ody.auth.service.apple;
 
 import com.ody.auth.config.AppleProperties;
+import com.ody.auth.service.SocialAuthUnlinkClient;
 import com.ody.common.exception.OdyServerErrorException;
 import com.ody.member.domain.AuthProvider;
 import com.ody.member.domain.ProviderType;
