+++ backend/src/main/java/com/ody/auth/service/apple/AppleValidateTokenClientErrorHandler.java
@@ -1,8 +1,7 @@
-package com.ody.auth.service;
+package com.ody.auth.service.apple;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ody.common.exception.OdyBadRequestException;
-import com.ody.common.exception.OdyServerErrorException;
 import java.io.IOException;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
@@ -14,9 +13,7 @@ import org.springframework.web.client.ResponseErrorHandler;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class KakaoAuthUnlinkClientErrorHandler implements ResponseErrorHandler {
-
-    private static final int NOT_REGISTERED_USER_CODE = -101;
+public class AppleValidateTokenClientErrorHandler implements ResponseErrorHandler {
 
     private final ObjectMapper objectMapper;
 
@@ -28,11 +25,6 @@ public class KakaoAuthUnlinkClientErrorHandler implements ResponseErrorHandler {
     @Override
     public void handleError(ClientHttpResponse response) throws IOException {
         Map responseBody = objectMapper.readValue(response.getBody(), Map.class);
-
-        log.info("카카오 연결 끊기를 실패했습니다. (응답: {})", responseBody);
-        if (responseBody.get("code").equals(NOT_REGISTERED_USER_CODE)) {
-            throw new OdyBadRequestException("카카오 연결 끊기 대상 유저가 아닙니다.");
-        }
-        throw new OdyServerErrorException("카카오 연결 끊기를 실패했습니다.");
+        throw new OdyBadRequestException("Apple Authorization Code 검증에 실패했습니다: " + responseBody.get("error_description").toString());
     }
 }
